moving the unpublish_* actions to the right place ( issue #277 )
