+++ example/blog/actions.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import division, absolute_import, unicode_literals
 
+from django.utils.translation import ugettext_lazy, pgettext_lazy
+from django.contrib import messages
+
 from djadmin2.actions import BaseListAction
 from djadmin2 import permissions
 
-from django.utils.translation import ugettext_lazy, pgettext_lazy
 
 
 class CustomPublishAction(BaseListAction):
@@ -46,3 +48,23 @@ class PublishAllItemsAction(BaseListAction):
 
     def process_queryset(self):
         self.get_queryset().update(published=True)
+
+
+def unpublish_items(request, queryset):
+    queryset.update(published=False)
+    messages.add_message(request, messages.INFO, ugettext_lazy(u'Items unpublished'))
+
+# Translators : action description
+unpublish_items.description = ugettext_lazy('Unpublish selected items')
+
+
+def unpublish_all_items(request, queryset):
+    queryset.update(published=False)
+    messages.add_message(
+        request,
+        messages.INFO,
+        ugettext_lazy('Items unpublished'),
+    )
+
+unpublish_all_items.description = ugettext_lazy('Unpublish all items')
+unpublish_all_items.only_selected = False
